+++ app/Werewolf/Options.hs
@@ -57,7 +57,6 @@ data Command
     | Poison Poison.Options
     | Protect Protect.Options
     | Quit
-    | Reveal
     | See See.Options
     | Start Start.Options
     | Status
@@ -108,7 +107,6 @@ werewolf = Options
         , command "poison"      $ info (helper <*> poison)      (fullDesc <> progDesc "Poison a player")
         , command "protect"     $ info (helper <*> protect)     (fullDesc <> progDesc "Protect a player")
         , command "quit"        $ info (helper <*> quit)        (fullDesc <> progDesc "Quit the current game")
-        , command "reveal"      $ info (helper <*> reveal)      (fullDesc <> progDesc "Reveal yourself")
         , command "see"         $ info (helper <*> see)         (fullDesc <> progDesc "See a player's allegiance")
         , command "start"       $ info (helper <*> start)       (fullDesc <> progDesc "Start a new game")
         , command "status"      $ info (helper <*> status)      (fullDesc <> progDesc "Get the status of the current game")
@@ -167,9 +165,6 @@ protect = Protect . Protect.Options <$> playerArgument
 
 quit :: Parser Command
 quit = pure Quit
-
-reveal :: Parser Command
-reveal = pure Reveal
 
 see :: Parser Command
 see = See . See.Options <$> playerArgument